Fixes #7 - Fully manage jetty.ini
This should ensure that the jetty.ini file is managed, and that
the (now commented out) exec that regenerates the keystore in older
puppetdb versions does not restart the puppetdb service in newer
versions.
If we add a facter fact for puppetdbversion, we could re-enable that
exec, based on the value of the fact.